user validate retrive,insert,update,delete

+++ src/main/java/controller/MaintenanceController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ws.rs.*;
@@ -7,6 +8,7 @@ import javax.ws.rs.core.MediaType;
 
 import com.google.gson.Gson;
 
+import client.ClientManager;
 import model.Interruption;
 import service.MaintenanceService;
 import service.MaintenencrServiceImpl;
@@ -14,15 +16,24 @@ import service.MaintenencrServiceImpl;
 @Path("/")
 public class MaintenanceController {
 	MaintenanceService maintenanceService = new MaintenencrServiceImpl();
+	ClientManager clientManager = new ClientManager();
 	Gson gson = new Gson();
 
 	@GET	
 	@Path("/Interruptions")
 	@Produces(MediaType.APPLICATION_JSON)
-	public String allInterruptions() {
-		List<Interruption> list = maintenanceService.allInterruptions();
-		String jsonString = gson.toJson(list);
-		return jsonString;
+	public String allInterruptions(@HeaderParam("Authorization") String auth) {
+		String currentUser = clientManager.getCurrentUser(auth);
+		if(currentUser.equals("login failed")) {
+			String output = "login failed";
+			String jsonString = gson.toJson(output);
+			return jsonString;
+		}else {
+			List<Interruption> list = maintenanceService.allInterruptions();
+			String jsonString = gson.toJson(list);
+			return jsonString;
+		}
+
 	}
 
 	@POST
@@ -30,11 +41,24 @@ public class MaintenanceController {
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	@Produces(MediaType.TEXT_PLAIN)
 
-	public String insertInterruption(@FormParam("intType") String intType,@FormParam("title") String title,@FormParam("description") String description,@FormParam("startDate") String startDate,@FormParam("endDate") String endDate,@FormParam("approval")String approval,@FormParam("custIDs") String custIDs) {
-		
-		String[] list = custIDs.split(";");
-		Interruption interruption = new Interruption(intType, title, description, startDate, endDate, list, approval);
-		String output = maintenanceService.insertInterruption(interruption);
+	public String insertInterruption(@FormParam("intType") String intType,@FormParam("title") String title,@FormParam("description") String description,@FormParam("startDate") String startDate,@FormParam("endDate") String endDate,@FormParam("approval")String approval,@FormParam("custIDs") String custIDs,@HeaderParam("Authorization") String auth) {
+
+		String[] list_str = custIDs.split(";");
+		List<String>list = new ArrayList<String>();
+		for (String string : list_str) {
+			list.add(string);
+		}
+		String output = "";
+		String currentUser = clientManager.getCurrentUser(auth);
+		System.out.println(currentUser);
+		if(currentUser.equals("login failed")) {
+			output = "login failed";
+		}
+		else {
+			Interruption interruption = new Interruption(intType, title, description, startDate, endDate, list, approval,currentUser);
+			output = maintenanceService.insertInterruption(interruption);
+		}
+
 		return output;
 	}
 
@@ -43,10 +67,19 @@ public class MaintenanceController {
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	@Produces(MediaType.TEXT_PLAIN)
 
-	public String updateInterruption(@FormParam("id") int id,@FormParam("intType") String intType,@FormParam("title") String title,@FormParam("description") String description,@FormParam("startDate") String startDate,@FormParam("endDate") String endDate,@FormParam("approval")String approval) {
-		Interruption interruption = new Interruption(intType, title, description, startDate, endDate, null, approval);
-		interruption.setId(id);
-		String output = maintenanceService.updateInterruption(interruption);
+	public String updateInterruption(@FormParam("id") int id,@FormParam("intType") String intType,@FormParam("title") String title,@FormParam("description") String description,@FormParam("startDate") String startDate,@FormParam("endDate") String endDate,@FormParam("approval")String approval,@HeaderParam("Authorization") String auth) {
+		String output = "";
+		String currentUser = clientManager.getCurrentUser(auth);
+		System.out.println(currentUser);
+		if(currentUser.equals("login failed")) {
+			output = "login failed";
+		}else {
+			Interruption interruption = new Interruption(intType, title, description, startDate, endDate, null, approval,currentUser);
+			interruption.setId(id);
+			output = maintenanceService.updateInterruption(interruption);	
+		}
+
+
 		return output;
 	}
 
@@ -55,19 +88,37 @@ public class MaintenanceController {
 	@Consumes(MediaType.APPLICATION_FORM_URLENCODED)
 	@Produces(MediaType.TEXT_PLAIN)
 
-	public String updateEffectedCustomer(@FormParam("id") int id,@FormParam("custIDs") String custIDs) {
-		String[] list = custIDs.split(";");
-		String output = maintenanceService.updateEffectedCustomer(id,list);
+	public String updateEffectedCustomer(@FormParam("id") int id,@FormParam("custIDs") String custIDs,@HeaderParam("Authorization") String auth) {
+
+		String output = "";
+		String currentUser = clientManager.getCurrentUser(auth);
+		System.out.println(currentUser);
+		if(currentUser.equals("login failed")) {
+			output = "login failed";
+		}else {
+			String[] list = custIDs.split(";");
+			output = maintenanceService.updateEffectedCustomer(id,list);
+		}
+
 		return output;
 	}
-	
-	
+
+
 	@DELETE
 	@Path("/Interruptions/{id}")
 	@Produces(MediaType.TEXT_PLAIN)
 
-	public String deleteInterruption(@PathParam("id") int id) {
-		String output = maintenanceService.deleteInterruption(id);
+	public String deleteInterruption(@PathParam("id") int id,@HeaderParam("Authorization") String auth) {
+
+		String output = "";
+		String currentUser = clientManager.getCurrentUser(auth);
+		System.out.println(currentUser);
+		if(currentUser.equals("login failed")) {
+			output = "login failed";
+		}else {
+			output = maintenanceService.deleteInterruption(id);
+		}
+
 		return output;
 	}
